backend/requirements.txt
Django~=6.0.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.11
pypdf>=4.0
reportlab>=4.0
django-cors-headers==4.9.0
python-dotenv==1.2.2
dj-database-url==3.1.2
gunicorn==23.0.0
whitenoise==6.9.0
requests==2.32.5
